Organized examples directory
