+++ examples/slp/type1/mint_token.js
@@ -1,4 +1,4 @@
-const Watchtower = require('../src')
+const Watchtower = require('../../../src')
 const watchtower = new Watchtower()
 
 const data = {
